+++ admin/drivers/mysql.inc.php
@@ -33,7 +33,7 @@ if (isset($_GET["mysql"])) {
 			public function open(string $server, string $username, string $password): bool
 			{
 				mysqli_report(MYSQLI_REPORT_OFF);
-				list($host, $port) = explode(":", $server, 2); // part after : is used for port or socket
+				list($host, $port) = host_port($server);
 
 				$key = Admin::get()->getConfig()->getSslKey();
 				$certificate = Admin::get()->getConfig()->getSslCertificate();
@@ -182,7 +182,8 @@ if (isset($_GET["mysql"])) {
 		{
 			public function open(string $server, string $username, string $password): bool
 			{
-				$dsn = "mysql:charset=utf8;host=" . str_replace(":", ";unix_socket=", preg_replace('~:(\d)~', ';port=\1', $server));
+				list($host, $port) = host_port($server);
+				$dsn = "mysql:charset=utf8;host=$host" . ($port ? (is_numeric($port) ? ";port=" : ";unix_socket=") . $port : "");
 
 				$options = [PDO::MYSQL_ATTR_LOCAL_INFILE => false];
 
@@ -947,7 +948,7 @@ ORDER BY ordinal_position";
 	* @param string
 	* @param string
 	* @param string
-	* @param numeric-string
+	* @param numeric-string|''
 	* @param ?array null means remove partitioning
 	* @return bool
 	*/
@@ -1171,7 +1172,8 @@ ORDER BY ordinal_position";
 		$aliases = ["bool", "boolean", "integer", "double precision", "real", "dec", "numeric", "fixed", "national char", "national varchar"];
 		$space = "(?:\\s|/\\*[\s\S]*?\\*/|(?:#|-- )[^\n]*\n?|--\r?\n)";
 		$enumLengthPattern = Driver::EnumLengthPattern;
-		$type_pattern = "((" . implode("|", array_merge(array_keys(Driver::get()->getTypes()), $aliases)) . ")\\b(?:\\s*\\(((?:[^'\")]|$enumLengthPattern)++)\\))?\\s*(zerofill\\s*)?(unsigned(?:\\s+zerofill)?)?)(?:\\s*(?:CHARSET|CHARACTER\\s+SET)\\s*['\"]?([^'\"\\s,]+)['\"]?)?";
+		$type_pattern = "((" . implode("|", array_merge(array_keys(Driver::get()->getTypes()), $aliases)) . ")\\b(?:\\s*\\(((?:[^'\")]|$enumLengthPattern)++)\\))?" .
+			"\\s*(zerofill\\s*)?(unsigned(?:\\s+zerofill)?)?)(?:\\s*(?:CHARSET|CHARACTER\\s+SET)\\s*['\"]?([^'\"\\s,]+)['\"]?)?(?:\\s*COLLATE\\s*['\"]?([^'\"\\s,]+)['\"]?)?"; // TODO store COLLATE
 		$inOut = implode("|", Driver::get()->getInOut());
 		$pattern = "$space*(" . ($type == "FUNCTION" ? "" : $inOut) . ")?\\s*(?:`((?:[^`]|``)*)`\\s*|\\b(\\S+)\\s+)$type_pattern";
 		$create = Connection::get()->getValue("SHOW CREATE $type " . idf_escape($name), 2);
@@ -1211,7 +1213,7 @@ ORDER BY ordinal_position";
 	* @return list<string[]> ["SPECIFIC_NAME" => , "ROUTINE_NAME" => , "ROUTINE_TYPE" => , "DTD_IDENTIFIER" => ]
 	*/
 	function routines() {
-		return get_rows("SELECT ROUTINE_NAME AS SPECIFIC_NAME, ROUTINE_NAME, ROUTINE_TYPE, DTD_IDENTIFIER, ROUTINE_COMMENT FROM information_schema.ROUTINES WHERE ROUTINE_SCHEMA = DATABASE()");
+		return get_rows("SELECT SPECIFIC_NAME, ROUTINE_NAME, ROUTINE_TYPE, DTD_IDENTIFIER, ROUTINE_COMMENT FROM information_schema.ROUTINES WHERE ROUTINE_SCHEMA = DATABASE()");
 	}
 
 	/** Get list of available routine languages
@@ -1299,12 +1301,31 @@ ORDER BY ordinal_position";
 		return "TRUNCATE " . table($table);
 	}
 
-	/** Get SQL command to change database
-	* @param string
-	* @return string
-	*/
-	function use_sql($database) {
-		return "USE " . idf_escape($database);
+	/**
+	 * Returns SQL command to create database.
+	 */
+	function create_database_sql(string $database, string $style = ""): string
+	{
+		$name = idf_escape($database);
+
+		$command = "";
+		if (str_contains($style, "CREATE") && ($create = Connection::get()->getValue("SHOW CREATE DATABASE $name", 1))) {
+			set_utf8mb4($create);
+			if ($style == "DROP+CREATE") {
+				$command = "DROP DATABASE IF EXISTS $name;\n";
+			}
+			$command .= "$create;\n";
+		}
+
+		return $command;
+	}
+
+	/**
+	 * Returns SQL command to change database.
+	 */
+	function use_sql(string $database, string $style = ""): string
+	{
+		return "USE " . idf_escape($database) . ";\n";
 	}
 
 	/**