+++ admin/scripts/functions.js
@@ -177,7 +177,7 @@ function trCheck(el) {
 function selectCount(id, count) {
 	const zero = count === 0 || count === '0' || count === '';
 
-	setHtml(id, (zero ? '' : '(' + (count + '').replace(/\B(?=(\d{3})+$)/g, thousandsSeparator) + ')'));
+	setHtml(id, '(' + (count + '').replace(/\B(?=(\d{3})+$)/g, thousandsSeparator) + ')');
 
 	const el = gid(id);
 	if (!el) return;
@@ -248,8 +248,8 @@ function formChecked(input, name) {
 function tableClick(event, click, canEdit = true) {
 	const td = parentTag(event.target, 'td');
 	let text;
-	if (canEdit && td && (text = td.getAttribute('data-text'))) {
-		if (selectClick.call(td, event, +text, td.getAttribute('data-warning'))) {
+	if (canEdit && td && (text = td.dataset.text)) {
+		if (selectClick.call(td, event, +text, td.dataset.warning)) {
 			return;
 		}
 	}
@@ -1291,7 +1291,7 @@ function getOffsetLeft(element) {
 
 oninput = event => {
 	const target = event.target;
-	const maxLength = target.getAttribute('data-maxlength');
+	const maxLength = target.dataset.maxlength;
 
 	// maxLength could be 0
 	target.classList.toggle('maxlength', target.value && maxLength != null && target.value.length > maxLength);