+++ admin/translations/cs.inc.php
@@ -126,6 +126,7 @@ return [
 	'File uploads are disabled.' => 'Nahrávání souborů není povoleno.',
 	'Unable to upload a file.' => 'Nepodařilo se nahrát soubor.',
 	'Maximum allowed file size is %sB.' => 'Maximální povolená velikost souboru je %sB.',
+	'The maximum number of files is %d. Select fewer files or increase the %s configuration directive.' => 'Maximální počet souborů je %d. Vyberte méně souborů nebo zvyšte hodnotu konfigurační direktivy %s.',
 	'Too big POST data. Reduce the data or increase the %s configuration directive.' => 'Příliš velká POST data. Zmenšete data nebo zvyšte hodnotu konfigurační direktivy %s.',
 	'You can upload a big SQL file via FTP and import it from server.' => 'Velký SQL soubor můžete nahrát pomocí FTP a importovat ho ze serveru.',
 	'File must be in UTF-8 encoding.' => 'Soubor musí být v kódování UTF-8.',